requirements.txt
alembic==1.7.7
altgraph==0.17.2
appdirs==1.4.4
attrs==21.4.0
banal==1.0.6
binaryornot==0.4.4
cachetools==5.3.1
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.3
cssselect==1.2.0
cssutils==2.7.1
et-xmlfile==1.1.0
future==0.18.3
greenlet==3.0.1
idna==3.7
jdcal==1.4.1
Jinja2==3.1.6
Levenshtein==0.22.0
lxml==4.9.3
macholib==1.16
Mako==1.2.2
MarkupSafe==2.1.3
openpyxl==3.0.9
packaging==23.2
pathspec==0.9.0
pbr==5.9.0
pefile==2022.5.30
Pillow==10.3.0
Pillow-PIL==0.1.dev0
platformdirs==2.5.2
PyQt6>=6.6.0
premailer==3.10.0
py3dns==4.0.0
pyodbc==5.0.1
python-barcode==0.15.1
python-dateutil==2.8.2
python-editor==1.0.4
python-Levenshtein==0.22.0
pytz==2023.3.post1
rapidfuzz==2.13.6
requests==2.32.2
six==1.16.0
SQLAlchemy==1.4.49
tendo==0.2.15
thefuzz==0.19.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.5
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2023.3
urllib3==2.2.2
validate-email==1.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-qt>=4.2.0
